Some optimisations for RTC Remove some inappropriate 16 bit ints Tweak project files to build in parallel Fix warnings from GCC 4.6-20110930

+++ mhvlib/MHV_io_ATmega1280.h
@@ -64,10 +64,10 @@
 #define MHV_AD_RESOLUTION		1024
 
 
-#define MHV_AD_REFERENCE_AREF	(0x00 << 6)
-#define MHV_AD_REFERENCE_AVCC	(0x01 << 6)
-#define MHV_AD_REFERENCE_1V1	(0x02 << 6)
-#define MHV_AD_REFERENCE_2V56	(0x03 << 6)
+#define MHV_AD_REFERENCE_AREF	(uint8_t)(0x00 << 6)
+#define MHV_AD_REFERENCE_AVCC	(uint8_t)(0x01 << 6)
+#define MHV_AD_REFERENCE_1V1	(uint8_t)(0x02 << 6)
+#define MHV_AD_REFERENCE_2V56	(uint8_t)(0x03 << 6)
 
 #define MHV_AD_CHANNEL_0			0x00
 #define MHV_AD_CHANNEL_1			0x01
