+++ mhvlib/MHV_io_ATmega168.h
@@ -53,9 +53,9 @@
 
 #define MHV_AD_RESOLUTION		1024
 
-#define MHV_AD_REFERENCE_AREF	(0x00 << 6)
-#define MHV_AD_REFERENCE_AVCC	(0x01 << 6)
-#define MHV_AD_REFERENCE_1V1	(0x03 << 6)
+#define MHV_AD_REFERENCE_AREF	(uint8_t)(0x00 << 6)
+#define MHV_AD_REFERENCE_AVCC	(uint8_t)(0x01 << 6)
+#define MHV_AD_REFERENCE_1V1	(uint8_t)(0x03 << 6)
 
 #define MHV_AD_CHANNEL_0			0x00
 #define MHV_AD_CHANNEL_1			0x01
